backend/requirements.txt
fastapi==0.128.8
uvicorn==0.40.0
anthropic==0.84.0
python-dotenv==1.0.1
starlette==0.52.1
aiofiles==23.2.1
